Modify the 6-size.c program using the data type integer

+++ hello_world/6-size.c
@@ -13,10 +13,10 @@ int main(void)
 	long long int ll;
 	float f;
 
-	printf("Size of a char: %lu byte(s)\n", (unsigned long) sizeof(c));
-	printf("Size of an int: %lu byte(s)\n", (unsigned long) sizeof(i));
-	printf("Size of a long int: %lu byte(s)\n", (unsigned long) sizeof(li));
-	printf("Size of a long long int: %lu byte(s)\n", (unsigned long) sizeof(ll));
-	printf("Size of a float: %lu byte(s)\n", (unsigned long) sizeof(f));
+	printf("Size of a char: %i byte(s)\n", sizeof(c));
+	printf("Size of an int: %i byte(s)\n", sizeof(i));
+	printf("Size of a long int: %i byte(s)\n", sizeof(li));
+	printf("Size of a long long int: %i byte(s)\n", sizeof(ll));
+	printf("Size of a float: %i byte(s)\n", sizeof(f));
 	return (0);
 }
